+++ src/decompress.rs
@@ -1,5 +1,5 @@
 use std::convert::TryInto as _;
-use raw::tjscalingfactor;
+use std::fmt;
 use crate::{Image, YuvImage, raw};
 use crate::common::{PixelFormat, Subsamp, Colorspace, Result, Error};
 use crate::handle::Handle;
@@ -29,62 +29,107 @@ pub struct DecompressHeader {
     pub subsamp: Subsamp,
     /// Colorspace of the compressed image.
     pub colorspace: Colorspace,
+    /// Is the image lossless JPEG?
+    pub is_lossless: bool,
 }
 
-/// Scaling factor for efficient scaling
+/// Fractional scaling factor.
 ///
-/// Turbojpeg gives us the ability to scaling by skipping DCT coefficients
+/// TurboJPEG can efficiently scale a JPEG image when decompressing. The scaling is implemented in
+/// the DCT algorithm, so scaling factors are limited to multiples of 1/8. Use
+/// [`Decompressor::supported_scaling_factors()`] to get the list of all scaling factors supported
+/// by the decompressor.
 #[derive(Debug, Copy, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum ScalingFactor {
-    /// No scaling
-    None,
-    /// 1/2 scale resolution
-    OneHalf,
-    /// 1/4 scale resolution
-    OneQuarter,
-    /// 1/8 scale resolution
-    OneEighth
+#[doc(alias = "tjscalingfactor")]
+pub struct ScalingFactor {
+    num: usize,
+    denom: usize,
 }
 
-impl Into<tjscalingfactor> for ScalingFactor {
-    fn into(self) -> tjscalingfactor {
-        match self {
-            Self::None => tjscalingfactor { num: 1, denom: 1 },
-            Self::OneHalf => tjscalingfactor { num: 1, denom: 2 },
-            Self::OneQuarter => tjscalingfactor { num: 1, denom: 4 },
-            Self::OneEighth => tjscalingfactor { num: 1, denom: 8 },
-        }
+impl ScalingFactor {
+    /// 1x scaling (no effect).
+    pub const ONE: Self = Self { num: 1, denom: 1 };
+    /// 1/2x scaling factor.
+    pub const ONE_HALF: Self = Self { num: 1, denom: 2 };
+    /// 1/4x scaling factor.
+    pub const ONE_QUARTER: Self = Self { num: 1, denom: 4 };
+    /// 1/8x scaling factor.
+    pub const ONE_EIGHTH: Self = Self { num: 1, denom: 8 };
+    /// 2x scaling factor.
+    pub const TWO: Self = Self { num: 2, denom: 1 };
+
+    /// Create a scaling factor from `num`-erator and `denom`-inator.
+    ///
+    /// We will simplify the fraction, so the numerator and denominator of the resulting fraction
+    /// might be different from what you pass.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// let s = turbojpeg::ScalingFactor::new(12, 8);
+    /// assert_eq!(s.num(), 3);
+    /// assert_eq!(s.denom(), 2);
+    /// ```
+    pub fn new(num: usize, denom: usize) -> Self {
+        let gcd = gcd::binary_usize(num, denom);
+        Self { num: num / gcd, denom: denom / gcd }
+    }
+
+    /// Get the numerator (the "3" in "3/4").
+    pub fn num(&self) -> usize {
+        self.num
+    }
+
+    /// Get the denominator (the "4" in "3/4").
+    pub fn denom(&self) -> usize {
+        self.denom
+    }
+
+    /// Compute the value of `dimension` scaled by this scaling factor.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// assert_eq!(turbojpeg::ScalingFactor::ONE_QUARTER.scale(400), 100);
+    /// assert_eq!(turbojpeg::ScalingFactor::ONE_QUARTER.scale(5), 2);
+    /// assert_eq!(turbojpeg::ScalingFactor::new(7, 8).scale(20), 18);
+    /// ```
+    #[doc(alias = "TJSCALED")]
+    pub fn scale(&self, dimension: usize) -> usize {
+        (dimension * self.num + self.denom - 1) / self.denom
+    }
+}
+
+impl fmt::Display for ScalingFactor {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}/{}", self.num, self.denom)
     }
 }
 
 impl DecompressHeader {
-    /// Convenience method that returns a new instance of the decompress header with the given scale
+    /// Scale the image size (width, height) in the header by the scaling factor.
+    ///
     /// # Example
     ///
     /// ```
-    /// // read JPEG data from file
+    /// // read JPEG header from file
     /// let jpeg_data = std::fs::read("examples/parrots.jpg")?;
-    ///
-    /// // read JPEG header to check initial dimensions
     /// let header = turbojpeg::read_header(&jpeg_data)?;
     /// assert_eq!((header.width, header.height), (384, 256));
     ///
-    ///
-    /// // check header size after scale
-    ///
-    /// let scale_factor = turbojpeg::ScalingFactor::OneHalf;
-    /// let scaled_header = header.with_scale(scale_factor);
+    /// // scale the header
+    /// let scale_factor = turbojpeg::ScalingFactor::ONE_HALF;
+    /// let scaled_header = header.scaled(scale_factor);
     /// assert_eq!((scaled_header.width, scaled_header.height), (192, 128));
     ///
     /// # Ok::<(), Box<dyn std::error::Error>>(())
-    pub fn with_scale(&self, factor: ScalingFactor) -> Self {
-        let factor: tjscalingfactor = factor.into();
+    /// ```
+    #[doc(alias = "TJSCALED")]
+    pub fn scaled(&self, factor: ScalingFactor) -> Self {
         Self {
-            width: (self.width + (factor.denom - 1) as usize) / factor.denom as usize,
-            height: (self.height + (factor.denom - 1) as usize) / factor.denom as usize,
-            subsamp: self.subsamp,
-            colorspace: self.colorspace
+            width: factor.scale(self.width),
+            height: factor.scale(self.height),
+            .. *self
         }
     }
 }
@@ -94,7 +139,7 @@ impl Decompressor {
     #[doc(alias = "tj3Init")]
     pub fn new() -> Result<Decompressor> {
         let handle = Handle::new(raw::TJINIT_TJINIT_DECOMPRESS)?;
-        Ok(Self { handle, scaling_factor: ScalingFactor::None })
+        Ok(Self { handle, scaling_factor: ScalingFactor::ONE })
     }
 
     /// Read the JPEG header without decompressing the image.
@@ -131,7 +176,60 @@ impl Decompressor {
             .try_into().map_err(|_| Error::IntegerOverflow("height"))?;
         let subsamp = Subsamp::from_int(self.handle.get(raw::TJPARAM_TJPARAM_SUBSAMP))?;
         let colorspace = Colorspace::from_int(self.handle.get(raw::TJPARAM_TJPARAM_COLORSPACE))?;
-        Ok(DecompressHeader { width, height, subsamp, colorspace })
+        let is_lossless = self.handle.get(raw::TJPARAM_TJPARAM_LOSSLESS) != 0;
+        Ok(DecompressHeader { width, height, subsamp, colorspace, is_lossless })
+    }
+
+    /// Set scaling factor for subsequent decompression operations.
+    ///
+    /// Only the scaling factors returned by
+    /// [`supported_scaling_factors()`][Self::supported_scaling_factors()] are supported, and a
+    /// scaling factor can only be used when decompressing lossy JPEG images.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// // read JPEG data from file
+    /// let jpeg_data = std::fs::read("examples/parrots.jpg")?;
+    ///
+    /// // initialize a decompressor with the scaling factor
+    /// let mut decompressor = turbojpeg::Decompressor::new()?;
+    /// let scaling = turbojpeg::ScalingFactor::ONE_HALF;
+    /// decompressor.set_scaling_factor(scaling);
+    ///
+    /// // read the JPEG header and downscale the width and height
+    /// let scaled_header = decompressor.read_header(&jpeg_data)?.scaled(scaling);
+    ///
+    /// // initialize the image (Image<Vec<u8>>)
+    /// let mut image = turbojpeg::Image {
+    ///     pixels: vec![0; 4 * scaled_header.width * scaled_header.height],
+    ///     width: scaled_header.width,
+    ///     pitch: 4 * scaled_header.width, // size of one image row in memory
+    ///     height: scaled_header.height,
+    ///     format: turbojpeg::PixelFormat::RGBA,
+    /// };
+    ///
+    /// // decompress the JPEG into the image
+    /// // (we use as_deref_mut() to convert from &mut Image<Vec<u8>> into Image<&mut [u8]>)
+    /// decompressor.decompress(&jpeg_data, image.as_deref_mut())?;
+    /// assert_eq!(&image.pixels[0..5], &[125, 121, 92, 255, 127]);
+    ///
+    /// # Ok::<(), Box<dyn std::error::Error>>(())
+    /// ```
+    #[doc(alias = "tj3SetScalingFactor")]
+    pub fn set_scaling_factor(&mut self, scaling_factor: ScalingFactor) -> Result<()> {
+        let num: libc::c_int = scaling_factor.num.try_into()
+            .map_err(|_| Error::IntegerOverflow("num"))?;
+        let denom: libc::c_int = scaling_factor.denom.try_into()
+            .map_err(|_| Error::IntegerOverflow("denom"))?;
+        self.handle.set_scaling_factor(raw::tjscalingfactor { num, denom })?;
+        self.scaling_factor = scaling_factor;
+        Ok(())
+    }
+
+    /// Get the scaling factor set by [`set_scaling_factor()`][Self::set_scaling_factor()].
+    pub fn scaling_factor(&self) -> ScalingFactor {
+        self.scaling_factor
     }
 
     /// Decompress a JPEG image in `jpeg_data` into `output`.
@@ -175,28 +273,15 @@ impl Decompressor {
         let width: libc::c_int = width.try_into().map_err(|_| Error::IntegerOverflow("width"))?;
         let pitch: libc::c_int = pitch.try_into().map_err(|_| Error::IntegerOverflow("pitch"))?;
         let height: libc::c_int = height.try_into().map_err(|_| Error::IntegerOverflow("height"))?;
+        let jpeg_data_len: raw::size_t = jpeg_data.len().try_into()
+            .map_err(|_| Error::IntegerOverflow("jpeg_data.len()"))?;
 
-        let res = unsafe {
-            raw::tj3DecompressHeader(
-                self.handle.as_ptr(),
-                jpeg_data.as_ptr(),
-                jpeg_data.len() as raw::size_t,
-            )
-        };
-        if res != 0 {
-            return Err(self.handle.get_error())
-        }
-
-        let scaled_header = self.read_header(jpeg_data)?.with_scale(self.scaling_factor);
-
-        if width < scaled_header.width as i32 || height < scaled_header.height as i32 {
-            return Err(Error::OutputTooSmall(scaled_header.width as i32, scaled_header.height as i32))
-        }
+        self.check_output_size(jpeg_data, width, height)?;
 
         let res = unsafe {
             raw::tj3Decompress8(
                 self.handle.as_ptr(),
-                jpeg_data.as_ptr(), jpeg_data.len() as raw::size_t,
+                jpeg_data.as_ptr(), jpeg_data_len,
                 pixels.as_mut_ptr(), pitch, format as i32,
             )
         };
@@ -204,50 +289,6 @@ impl Decompressor {
             return Err(self.handle.get_error())
         }
 
-        Ok(())
-    }
-
-
-    /// Set scaling factor to take effect when decompressing
-    /// Note that this will not work with lossless jpeg images
-    // ///
-    // /// # Example
-    // ///
-    // /// ```
-    // /// // read JPEG data from file
-    // /// let jpeg_data = std::fs::read("examples/parrots.jpg")?;
-    // ///
-    // /// // define a scaling
-    // /// let scaling = turbojpeg::ScalingFactor::OneHalf;
-    // ///
-    // /// // initialize a decompressor with the scaling factor
-    // /// let mut decompressor = turbojpeg::Decompressor::new()?;
-    // /// decompressor.set_scaling_factor(scaling);
-    // ///
-    // /// 
-    // ///
-    // /// // read the JPEG header, downscaling the width and height
-    // /// let scaled_header = decompressor.read_header(&jpeg_data)?.with_scale(scaling);
-    // ///
-    // /// // initialize the image (Image<Vec<u8>>)
-    // /// let mut image = turbojpeg::Image {
-    // ///     pixels: vec![0; 4 * scaled_header.width * scaled_header.height],
-    // ///     width: scaled_header.width,
-    // ///     pitch: 4 * scaled_header.width, // size of one image row in memory
-    // ///     height: scaled_header.height,
-    // ///     format: turbojpeg::PixelFormat::RGBA,
-    // /// };
-    // ///
-    // /// // decompress the JPEG into the image
-    // /// // (we use as_deref_mut() to convert from &mut Image<Vec<u8>> into Image<&mut [u8]>)
-    // /// decompressor.decompress_with_scaling(&jpeg_data, image.as_deref_mut(), scaling)?;
-    // /// assert_eq!(&image.pixels[0..5], &[125, 121, 92, 255, 127]);
-    // ///
-    // /// # Ok::<(), Box<dyn std::error::Error>>(())
-    // /// ```
-    pub fn set_scaling_factor(&mut self, scaling_factor: ScalingFactor) -> Result<()> {
-        self.scaling_factor = scaling_factor;
-        self.handle.set_scaling_factor(scaling_factor.into())?;
         Ok(())
     }
 
@@ -293,16 +334,12 @@ impl Decompressor {
         output.assert_valid(output.pixels.len());
         let YuvImage { pixels, width, align, height, subsamp: _ } = output;
         let width: libc::c_int = width.try_into().map_err(|_| Error::IntegerOverflow("width"))?;
-        let align = align.try_into().map_err(|_| Error::IntegerOverflow("align"))?;
+        let align: libc::c_int = align.try_into().map_err(|_| Error::IntegerOverflow("align"))?;
         let height: libc::c_int = height.try_into().map_err(|_| Error::IntegerOverflow("height"))?;
-        let jpeg_data_len = jpeg_data.len().try_into()
+        let jpeg_data_len: raw::size_t = jpeg_data.len().try_into()
             .map_err(|_| Error::IntegerOverflow("jpeg_data.len()"))?;
 
-        let scaled_header = self.read_header(jpeg_data)?.with_scale(self.scaling_factor);
-
-        if width < scaled_header.width as i32 || height < scaled_header.height as i32 {
-            return Err(Error::OutputTooSmall(scaled_header.width as i32, scaled_header.height as i32))
-        }
+        self.check_output_size(jpeg_data, width, height)?;
 
         let res = unsafe {
             raw::tj3DecompressToYUV8(
@@ -316,6 +353,58 @@ impl Decompressor {
         }
 
         Ok(())
+    }
+
+    fn check_output_size(&mut self, jpeg_data: &[u8], width: libc::c_int, height: libc::c_int) -> Result<()> {
+        let header = self.read_header(jpeg_data)?;
+
+        if header.is_lossless && self.scaling_factor != ScalingFactor::ONE {
+            return Err(Error::CannotScaleLossless)
+        }
+        let scaled_width = self.scaling_factor.scale(header.width);
+        let scaled_height = self.scaling_factor.scale(header.height);
+
+        if width < scaled_width as i32 || height < scaled_height as i32 {
+            return Err(Error::OutputTooSmall(scaled_width as i32, scaled_height as i32))
+        }
+
+        Ok(())
+    }
+
+    /// Get the list of scaling factors supported for decompression.
+    ///
+    /// At the time of this writing, TurboJPEG supports all multiples of 1/8 between 1/8 and 2 as
+    /// scaling factors, but it's unclear whether this is guaranteed to continue to be the case in
+    /// all future versions.
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// let factors = turbojpeg::Decompressor::supported_scaling_factors();
+    /// for num in 1..16 {
+    ///     let multiple_of_8 = turbojpeg::ScalingFactor::new(num, 8);
+    ///     assert!(factors.iter().find(|&f| *f == multiple_of_8).is_some());
+    /// }
+    /// ```
+    #[doc(alias = "tj3GetScalingFactors")]
+    pub fn supported_scaling_factors() -> Vec<ScalingFactor> {
+        let mut count: libc::c_int = 0;
+        let ptr: *const raw::tjscalingfactor = unsafe {
+            raw::tj3GetScalingFactors(&mut count as *mut _)
+        };
+        let count: usize = count.try_into()
+            .expect("tj3GetScalingFactors() returned a number that cannot be converted to usize");
+
+        let mut list = Vec::with_capacity(count);
+        for i in 0..count {
+            let factor = unsafe { ptr.add(i).read() };
+            let num: usize = factor.num.try_into()
+                .expect("Numerator of a tjscalingfactor cannot be converted to usize");
+            let denom: usize = factor.denom.try_into()
+                .expect("Denominator of a tjscalingfactor cannot be converted to usize");
+            list.push(ScalingFactor { num, denom });
+        }
+        list
     }
 }